third_party/WebKit/LayoutTests/fast/css/background-image-size-serialization.html

<script src="../../resources/js-test.js"></script>
<style>
.noposition
{
    background: url("foo.png");
    background-size: contain;
}
.position1
{
    background: url("foo.png");
    background-position: 1px;
    background-size: contain;
}
.position2
{
    background: url("foo.png");
    background-position-x: 1px;
    background-size: contain;
}
.position3
{
    background: url("foo.png");
    background-position: 1px 1px;
    background-size: contain;
}
</style>
<script>
var cssRules = document.styleSheets[1].cssRules;

shouldBe('cssRules.length', '4');
shouldBeEqualToString('cssRules[0].cssText', '.noposition { background: url("foo.png") 0% 0% / contain; }');
shouldBeEqualToString('cssRules[1].cssText', '.position1 { background: url("foo.png") 1px center / contain; }');
shouldBeEqualToString('cssRules[2].cssText', '.position2 { background: url("foo.png") 1px / contain; }');
shouldBeEqualToString('cssRules[3].cssText', '.position3 { background: url("foo.png") 1px 1px / contain; }');
</script>
